4-3-20 (a)
Added a bunch of new configs for proababilities of items and attachments.

Broke blacklists down into categories for ease of use.

Added blacklist checks for loot and gear tables that are not based on dynamicConfigs.

Added classname checks for loot and gear throughout so all are checked now.

Added a new config, player rewards.

Moved some recurring code at the end of configs to new .sqf in the init folder to consolidate this and make future updates easier. No more having ot change things in multiple files.

Updated some thoughts and todos